# api/hqbase-mail-api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HQBase Mail API",
    "version": "1.0.0",
    "description": "Stable v1 mail API for HQBase clients, automations, and agents. Additive fields may be introduced within v1; clients must ignore unknown response fields. Administrative APIs are not part of this contract."
  },
  "servers": [
    {
      "url": "{origin}",
      "variables": {
        "origin": {
          "default": "https://mail.example.com",
          "description": "HQBase origin"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/v1/messages": {
      "get": {
        "summary": "List or search messages",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "One page of messages, newest activity first.",
            "headers": {
              "Link": {
                "required": false,
                "description": "RFC 8288 link to the next page, as `<url>; rel=\"next\"`. The header is absent on the last page.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. `INVALID_LIMIT` when limit is not an integer from 1 to 100. `INVALID_CURSOR` when the cursor is malformed or is not a message cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "listMessages",
        "parameters": [
          {
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["inbox", "sent", "drafts", "archived", "trash", "catchall"]
            },
            "description": "Folder to list."
          },
          {
            "in": "query",
            "name": "mailboxId",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Limit results to one accessible mailbox."
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "Search subject, sender, recipient, and message text."
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of messages in the page."
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor from the previous page. Read it from the Link header and send it unchanged."
          }
        ]
      }
    },
    "/api/v1/messages/{id}": {
      "get": {
        "summary": "Get a message",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "getMessage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ]
      }
    },
    "/api/v1/messages/{id}/thread": {
      "get": {
        "summary": "Get a message thread",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Accessible thread messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDetail"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "getMessageThread",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ]
      }
    },
    "/api/v1/messages/{id}/html": {
      "get": {
        "summary": "Get sanitized message HTML",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sanitized HTML",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageHtml"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "getMessageHtml",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "in": "query",
            "name": "loadRemoteImages",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [1]
            },
            "description": "Load remote images for this response only."
          }
        ]
      }
    },
    "/api/v1/messages/{id}/inline/{attachmentId}": {
      "get": {
        "summary": "Render a safe inline image",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inline image",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "contentEncoding": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported inline media",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "getInlineAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ]
      }
    },
...
    "/api/v1/messages/{id}/remote-media/trust": {
      "post": {
        "summary": "Trust remote media from this sender",
        "security": [
          {
            "oauth2": ["mail:write"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Trust preference",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteMediaTrust"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "trustRemoteMedia",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ]
      }
    },
    "/api/v1/messages/{id}/{action}": {
      "post": {
        "summary": "Update message state",
        "security": [
          {
            "oauth2": ["mail:write"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "updateMessage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["read", "unread", "star", "unstar", "archive", "trash", "restore"]
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "Audience-bound OAuth token. Request resource={origin}/api/v1.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/auth/oauth2/authorize",
            "tokenUrl": "/api/auth/oauth2/token",
            "refreshUrl": "/api/auth/oauth2/token",
            "scopes": {
              "mail:read": "Read accessible mail",
              "mail:write": "Organize accessible mail",
              "mail:send": "Manage drafts and send mail",
              "offline_access": "Request a refresh token"
            }
          }
        }
      },
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Same-origin HQBase web session. Secure deployments may prefix the cookie name."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "MessageSummary": {
        "type": "object",
        "required": [
          "id",
          "threadId",
          "mailboxId",
          "direction",
          "folder",
          "fromAddress",
          "to",
          "subject",
          "snippet",
          "receivedAt",
          "sentAt",
          "readAt",
          "starredAt",
          "hasAttachments",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "threadId": {
            "type": "string"
          },
          "mailboxId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "direction": {
            "type": "string",
            "enum": ["inbound", "outbound"]
          },
          "folder": {
            "type": "string",
            "enum": ["inbox", "sent", "drafts", "archived", "trash", "catchall"]
          },
          "fromAddress": {
            "type": "string",
            "format": "email"
          },
          "to": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subject": {
            "type": "string"
          },
          "snippet": {
            "type": "string"
          },
          "receivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "sentAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "readAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "starredAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "hasAttachments": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Attachment": {
        "type": "object",
        "required": [
          "id",
          "messageId",
          "filename",
          "contentType",
          "sizeBytes",
          "contentId",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer"
          },
          "contentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MessageDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MessageSummary"
          },
          {
            "type": "object",
            "required": [
              "cc",
              "bcc",
              "deliveredToAddress",
              "textBody",
              "htmlAvailable",
              "messageId",
              "inReplyTo",
              "references",
              "attachments"
            ],
            "properties": {
              "cc": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "bcc": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "deliveredToAddress": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "textBody": {
                "type": "string"
              },
              "htmlAvailable": {
                "type": "boolean"
              },
              "messageId": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "inReplyTo": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "references": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "attachments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            }
          }
        ]
      },
      "MessageHtml": {
        "type": "object",
        "required": ["hasRemoteImages", "html", "quotedHtml", "remoteMediaTrusted"],
        "properties": {
          "hasRemoteImages": {
            "type": "boolean"
          },
          "html": {
            "type": "string"
          },
          "quotedHtml": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "remoteMediaTrusted": {
            "type": "boolean"
          }
        }
      },
      "RemoteMediaTrust": {
        "type": "object",
        "required": ["remoteMediaTrusted"],
        "properties": {
          "remoteMediaTrusted": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
    }
  }
}